+++ src/runtime/attachment_manager/index.ts
@@ -11,19 +11,30 @@ export interface IAttachmentMedia {
 export interface IAttachment {
   data: string
   media: IAttachmentMedia
+  fileName?: string
 }
 
 export type IAttachFunction = (attachment: IAttachment) => void
 
-export type ICreateStringAttachment = (data: string, mediaType?: string) => void
-export type ICreateBufferAttachment = (data: Buffer, mediaType: string) => void
+export interface ICreateAttachmentOptions {
+  mediaType: string
+  fileName?: string
+}
+export type ICreateStringAttachment = (
+  data: string,
+  mediaTypeOrOptions?: string | ICreateAttachmentOptions
+) => void
+export type ICreateBufferAttachment = (
+  data: Buffer,
+  mediaTypeOrOptions: string | ICreateAttachmentOptions
+) => void
 export type ICreateStreamAttachment = (
   data: Readable,
-  mediaType: string
+  mediaTypeOrOptions: string | ICreateAttachmentOptions
 ) => Promise<void>
 export type ICreateStreamAttachmentWithCallback = (
   data: Readable,
-  mediaType: string,
+  mediaTypeOrOptions: string | ICreateAttachmentOptions,
   callback: () => void
 ) => void
 export type ICreateAttachment = ICreateStringAttachment &
@@ -45,33 +56,47 @@ export default class AttachmentManager {
 
   create(
     data: Buffer | Readable | string,
-    mediaType?: string,
+    mediaTypeOrOptions?: string | ICreateAttachmentOptions,
     callback?: () => void
   ): void | Promise<void> {
+    const options = normaliseOptions(mediaTypeOrOptions)
     if (Buffer.isBuffer(data)) {
-      if (doesNotHaveValue(mediaType)) {
+      if (doesNotHaveValue(options.mediaType)) {
         throw Error('Buffer attachments must specify a media type')
       }
-      this.createBufferAttachment(data, mediaType)
+      this.createBufferAttachment(data, options.mediaType, options.fileName)
     } else if (isStream.readable(data)) {
-      if (doesNotHaveValue(mediaType)) {
+      if (doesNotHaveValue(options.mediaType)) {
         throw Error('Stream attachments must specify a media type')
       }
-      return this.createStreamAttachment(data, mediaType, callback)
+      return this.createStreamAttachment(
+        data,
+        options.mediaType,
+        options.fileName,
+        callback
+      )
     } else if (typeof data === 'string') {
-      if (doesNotHaveValue(mediaType)) {
-        mediaType = 'text/plain'
+      if (doesNotHaveValue(options.mediaType)) {
+        options.mediaType = 'text/plain'
       }
-      if (mediaType.startsWith('base64:')) {
-        this.createStringAttachment(data, {
-          encoding: messages.AttachmentContentEncoding.BASE64,
-          contentType: mediaType.replace('base64:', ''),
-        })
+      if (options.mediaType.startsWith('base64:')) {
+        this.createStringAttachment(
+          data,
+          {
+            encoding: messages.AttachmentContentEncoding.BASE64,
+            contentType: options.mediaType.replace('base64:', ''),
+          },
+          options.fileName
+        )
       } else {
-        this.createStringAttachment(data, {
-          encoding: messages.AttachmentContentEncoding.IDENTITY,
-          contentType: mediaType,
-        })
+        this.createStringAttachment(
+          data,
+          {
+            encoding: messages.AttachmentContentEncoding.IDENTITY,
+            contentType: options.mediaType,
+          },
+          options.fileName
+        )
       }
     } else {
       throw Error(
@@ -80,17 +105,26 @@ export default class AttachmentManager {
     }
   }
 
-  createBufferAttachment(data: Buffer, mediaType: string): void {
-    this.createStringAttachment(data.toString('base64'), {
-      encoding: messages.AttachmentContentEncoding.BASE64,
-      contentType: mediaType,
-    })
+  createBufferAttachment(
+    data: Buffer,
+    mediaType: string,
+    fileName?: string
+  ): void {
+    this.createStringAttachment(
+      data.toString('base64'),
+      {
+        encoding: messages.AttachmentContentEncoding.BASE64,
+        contentType: mediaType,
+      },
+      fileName
+    )
   }
 
   createStreamAttachment(
     data: Readable,
     mediaType: string,
-    callback: () => void
+    fileName?: string,
+    callback?: () => void
   ): void | Promise<void> {
     const promise = new Promise<void>((resolve, reject) => {
       const buffers: Uint8Array[] = []
@@ -98,7 +132,7 @@ export default class AttachmentManager {
         buffers.push(chunk)
       })
       data.on('end', () => {
-        this.createBufferAttachment(Buffer.concat(buffers), mediaType)
+        this.createBufferAttachment(Buffer.concat(buffers), mediaType, fileName)
         resolve()
       })
       data.on('error', reject)
@@ -110,7 +144,29 @@ export default class AttachmentManager {
     }
   }
 
-  createStringAttachment(data: string, media: IAttachmentMedia): void {
-    this.onAttachment({ data, media })
+  createStringAttachment(
+    data: string,
+    media: IAttachmentMedia,
+    fileName?: string
+  ): void {
+    this.onAttachment({
+      data,
+      media,
+      ...(fileName ? { fileName } : {}),
+    })
   }
+}
+
+function normaliseOptions(
+  mediaTypeOrOptions?: string | ICreateAttachmentOptions
+): Partial<ICreateAttachmentOptions> {
+  if (!mediaTypeOrOptions) {
+    return {}
+  }
+  if (typeof mediaTypeOrOptions === 'string') {
+    return {
+      mediaType: mediaTypeOrOptions,
+    }
+  }
+  return mediaTypeOrOptions
 }